+++ src/helper/TrainingImageReader.java
@@ -78,7 +78,7 @@ public class TrainingImageReader{
 		// System.out.println(label);
 		labelOffset++;
 
-		double[] img = new double[width * height];
+		float[] img = new float[width * height];
 
 		for(int i = 0; i < width; i++)
 			for(int j = 0; j < height; j++){
