+++ src/recognition/TangentDistance.java
@@ -7,12 +7,12 @@ import transformations.TransformationType;
 
 class TangentDistance{
 
-	static double calculateTangentDistance(ImagePoint img1, ImagePoint img2){
+	static float calculateTangentDistance(ImagePoint img1, ImagePoint img2){
 
-		double[][] tangentsE = 
+		float[][] tangentsE = 
 				img1.getTangents();
 //		 calculateTangents( img1.getImageArray(), 28, 28, 0 );
-		double[][] tangentsP = 
+		float[][] tangentsP = 
 				img2.getTangents();
 //		 calculateTangents( img2.getImageArray(), 28, 28, 0 );
 
@@ -40,25 +40,25 @@ class TangentDistance{
 
 		Matrix E2 = E.plus( Le.times( alphaE ) ), P2 = P.plus( Lp.times( alphaP ) );
 
-		double dist = E2.minus( P2 ).norm2();
+		float dist = E2.minus( P2 ).norm2();
 
 		// System.out.println( "Distance = " + dist );
 
 		return dist;
 	}
 
-	private static final double TEMPLATE_FACTOR_1 = 0.1667, TEMPLATE_FACTOR_2 = 0.6667,
-			TEMPLATE_FACTOR_3 = 0.08;
-	private static final double ADDITIVE_BRIGHTNESS_VALUE = 0.1;
+	private static final float TEMPLATE_FACTOR_1 = 0.1667f, TEMPLATE_FACTOR_2 = 0.6667f,
+			TEMPLATE_FACTOR_3 = 0.08f;
+	private static final float ADDITIVE_BRIGHTNESS_VALUE = 0.1f;
 
 	// Two dimensional access on images saved in one dimensional array
 	private static int tdIndex(int y, int x, int width){
 		return y * width + x;
 	}
 
-	static double[][] calculateTangents(double[] image, int height, int width, double bg){
+	static float[][] calculateTangents(float[] image, int height, int width, float bg){
 
-		double[][] tangents = new double[3][height * width];
+		float[][] tangents = new float[3][height * width];
 
 		tangents[0] = calculateTangent( image, height, width, TransformationType.TranslationX, bg );
 		tangents[1] = calculateTangent( image, height, width, TransformationType.TranslationY, bg );
@@ -71,45 +71,45 @@ class TangentDistance{
 		return tangents;
 	}
 
-	private static double[] calculateTangent(double[] image, int height, int width,
-			TransformationType type, double background){
+	private static float[] calculateTangent(float[] image, int height, int width,
+			TransformationType type, float background){
 		int j, k, ind, maxdim;
-		double tp, factorW, offsetW, factorH, factor, offsetH, halfbg;
-		double[] tmp, x1, x2;
+		float tp, factorW, offsetW, factorH, factor, offsetH, halfbg;
+		float[] tmp, x1, x2;
 
 		int size = height * width;
 		maxdim = ( height > width ) ? height : width;
 
-		tmp = new double[maxdim];
-		x1 = new double[size];
-		x2 = new double[size];
+		tmp = new float[maxdim];
+		x1 = new float[size];
+		x2 = new float[size];
 
-		factorW = ( (double) width * 0.5 );
-		offsetW = 0.5 - factorW;
-		factorW = 1.0 / factorW;
+		factorW = ( (float) width * 0.5f );
+		offsetW = 0.5f - factorW;
+		factorW = 1.0f / factorW;
 
-		factorH = ( (double) height * 0.5 );
-		offsetH = 0.5 - factorH;
-		factorH = 1.0 / factorH;
+		factorH = ( (float) height * 0.5f );
+		offsetH = 0.5f - factorH;
+		factorH = 1.0f / factorH;
 
 		factor = ( factorH < factorW ) ? factorH : factorW; // min
 
-		halfbg = 0.5 * background;
+		halfbg = 0.5f * background;
 
 		/* x1 shift along width */
 		/* first use mask 1 0 -1 */
 		for(k = 0; k < height; k++){
 			/* first column */
 			ind = tdIndex( k, 0, width );
-			x1[ind] = halfbg - image[ind + 1] * 0.5;
+			x1[ind] = halfbg - image[ind + 1] * 0.5f;
 			/* other columns */
 			for(j = 1; j < width - 1; j++){
 				ind = tdIndex( k, j, width );
-				x1[ind] = ( image[ind - 1] - image[ind + 1] ) * 0.5;
+				x1[ind] = ( image[ind - 1] - image[ind + 1] ) * 0.5f;
 			}
 			/* last column */
 			ind = tdIndex( k, width - 1, width );
-			x1[ind] = image[ind - 1] * 0.5 - halfbg;
+			x1[ind] = image[ind - 1] * 0.5f - halfbg;
 		}
 		/* now compute 3x3 template */
 		/* first line */
@@ -160,15 +160,15 @@ class TangentDistance{
 		/* first use mask 1 0 -1 */
 		for(j = 0; j < width; j++){
 			/* first line */
-			x2[j] = halfbg - image[j + width] * 0.5;
+			x2[j] = halfbg - image[j + width] * 0.5f;
 			/* other lines */
 			for(k = 1; k < height - 1; k++){
 				ind = tdIndex( k, j, width );
-				x2[ind] = ( image[ind - width] - image[ind + width] ) * 0.5;
+				x2[ind] = ( image[ind - width] - image[ind + width] ) * 0.5f;
 			}
 			/* last line */
 			ind = tdIndex( height - 1, j, width );
-			x2[ind] = image[ind - width] * 0.5 - halfbg;
+			x2[ind] = image[ind - width] * 0.5f - halfbg;
 		}
 
 		/* now compute 3x3 template */
@@ -217,7 +217,7 @@ class TangentDistance{
 
 		/* now go through the tangents */
 
-		double[] tangent = new double[size];
+		float[] tangent = new float[size];
 
 		switch (type){
 			case TranslationX:
